Refactor routes and update dotenv dependency declared
