+++ routes/user.js
@@ -1,3 +1,5 @@
+require("dotenv").config();
+
 var express = require("express");
 var router = express.Router();
 const UserModel = require("../models/UserModel");
